TGP-377: Refactoring
