+++ test/uk/gov/hmrc/tradergoodsprofiles/models/GetRecordResponseSpec.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.tradergoodsprofiles.models
 
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.{JsObject, JsResult, Json}
-import uk.gov.hmrc.tradergoodsprofiles.controllers.support.response.GetRecordResponseSupport
+import uk.gov.hmrc.tradergoodsprofiles.controllers.support.responses.GetRecordResponseSupport
 
 import java.time.Instant
